Remove API namespace - Move main classes to the root namespace - Move TestMiddleware.php to a separate Tests\Middleware namespace
