+++ src/RetryCallback.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rapkis\Controld\Api;
+namespace Rapkis\Controld;
 
 use Illuminate\Http\Client\ConnectionException;
 use Throwable;
